% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_read.R
\name{df_read}
\alias{df_read}
\title{Read data frame}
\usage{
df_read(x)
}
\arguments{
\item{file_name}{a path to read data frame from the file}

\item{comp_file}{path to save the companion file}
}
\description{
\code{df_read()} functions to read data frames from plain text delimited files while retaining factor levels
}
\seealso{
\code{readr::read.delim())}
}
