% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_read_write.R
\name{df_write}
\alias{df_write}
\alias{df_read}
\title{Read and Write data frame}
\usage{
df_write(x,file_name, comp_file)

df_read(file_name, comp_file = "file_level.txt")
}
\arguments{
\item{x}{a data frame}

\item{file_name}{a path to save the data frame to the directory or read data frame from the file}

\item{comp_file}{path to save the companion file}
}
\value{
\code{df_read()} returns a data frame read from  csv files.
An error is raised if the input is not a dataframe,or there exits no such file.
}
\description{
\code{df_read()} functions to read data frames to plain text delimited files while retaining factor levels
}
\seealso{
\code{readr::read_csv()}, \code{readr::write_csv()}
}
