% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_write.R
\name{df_write}
\alias{df_write}
\title{Write data frame}
\usage{
df_write(df,filename)
}
\arguments{
\item{file_name}{a path to write data frame to the file}
}
\description{
\code{df_read()} functions to write data frames to plain text delimited files while retaining factor levels
}
\examples{
 df_write(c(1,2,3),"numbers.txt")
}
\seealso{
\code{write.table}
}
