% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_check.R
\name{factor_check}
\alias{factor_check}
\title{To test whether a given argument is a factor or not}
\usage{
factor_check(x)
}
\arguments{
\item{x}{factor}
}
\value{
Return nothing if it is a factor, or else a message stating "This is not a factor, please ensure the adequate data type" and also stop the function.
}
\description{
To test whether a given argument is a factor or not
}
\examples{
factor_check(iris$Species)
}
