% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freorder.R
\name{freorder}
\alias{freorder}
\title{Reorder Levels of a Factor}
\usage{
freorder(x)
}
\arguments{
\item{x}{a factor, an atomic vector. The vector is treated as a categorical variable whose levels will be reordered.}
}
\value{
By default, the function will return the factor in the descending order.
}
\description{
Reorder Levels of a Factor
}
\seealso{
\link{reorder}
freorder(factor(c(1000,100,10)))\n
freorder(factor(c("a","z","m")))
}
