% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fset_level.R
\name{fset_level}
\alias{fset_level}
\title{Set Levels of Factor}
\usage{
fset_level(x)
}
\arguments{
\item{x}{factor}
}
\value{
Return a factor with its level identical to the order of the data
}
\description{
sets levels to the order in which they appear in the data, i.e. set the levels “as is”
}
\examples{
fset_level(factor(c("banana","coconut","apple")))
this will return a sorted level: apple, babana, coconut in the alphabetical order.
}
\seealso{
\link{base::factor}
}
